Telegraf private linter
